requirements.txt
SQLAlchemy==1.4.22
mysql-connector-python==8.0.26
Faker==9.8.1
python-dotenv==0.19.0